refactor: 1. Lottery에서 필요없는 인스턴스 변수 matched 제거 후 match 결과 바로 반환하도록 분리 2. isEqual 메소드 네이밍 변경
